feat: move `filename` option to generator/minimizer
