--- test/loader-filename-option.test.js
@@ -1,171 +1,0 @@
-import path from "path";
-
-import fileType from "file-type";
-
-import { fixturesPath, runWebpack, clearDirectory } from "./helpers";
-import ImageMinimizerPlugin from "../src/index.js";
-
-describe('loader "filename" option', () => {
-  beforeAll(() => clearDirectory(path.resolve(__dirname, "outputs")));
-  afterAll(() => clearDirectory(path.resolve(__dirname, "outputs")));
-
-  it("should emit new transformed to webp asset with flat filename", async () => {
-    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-1");
-    const stats = await runWebpack({
-      entry: path.join(fixturesPath, "./loader-single.js"),
-      imageminLoaderOptions: {
-        filename: "[name].webp",
-        generator: [
-          {
-            preset: "webp",
-            implementation: ImageMinimizerPlugin.squooshGenerate,
-            options: {
-              encodeOptions: {
-                webp: {
-                  lossless: 1,
-                },
-              },
-            },
-          },
-        ],
-      },
-      output: {
-        path: outputDir,
-      },
-    });
-    const { compilation } = stats;
-    const { warnings, errors } = compilation;
-
-    const transformedAsset = path.resolve(
-      __dirname,
-      outputDir,
-      "loader-test.webp"
-    );
-
-    const transformedExt = await fileType.fromFile(transformedAsset);
-
-    expect(/image\/webp/i.test(transformedExt.mime)).toBe(true);
-    expect(warnings).toHaveLength(0);
-    expect(errors).toHaveLength(0);
-  });
-
-  it("should emit new transformed to webp asset with nested filename", async () => {
-    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-2");
-    const stats = await runWebpack({
-      entry: path.join(fixturesPath, "./loader-single.js"),
-      imageminLoaderOptions: {
-        filename: "deep/[path][name].webp",
-        generator: [
-          {
-            preset: "webp",
-            implementation: ImageMinimizerPlugin.squooshGenerate,
-            options: {
-              encodeOptions: {
-                webp: {
-                  lossless: 1,
-                },
-              },
-            },
-          },
-        ],
-      },
-      output: {
-        path: outputDir,
-      },
-    });
-    const { compilation } = stats;
-    const { warnings, errors } = compilation;
-
-    const transformedAsset = path.resolve(
-      __dirname,
-      outputDir,
-      "deep/nested/deep/loader-test.webp"
-    );
-
-    const transformedExt = await fileType.fromFile(transformedAsset);
-
-    expect(/image\/webp/i.test(transformedExt.mime)).toBe(true);
-    expect(warnings).toHaveLength(0);
-    expect(errors).toHaveLength(0);
-  });
-
-  it("should emit new transformed to webp asset with filename  pointing to other directory", async () => {
-    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-3");
-    const stats = await runWebpack({
-      entry: path.join(fixturesPath, "./loader-single.js"),
-      imageminLoaderOptions: {
-        filename: "other/[name].webp",
-        generator: [
-          {
-            preset: "webp",
-            implementation: ImageMinimizerPlugin.squooshGenerate,
-            options: {
-              encodeOptions: {
-                webp: {
-                  lossless: 1,
-                },
-              },
-            },
-          },
-        ],
-      },
-      output: {
-        path: outputDir,
-      },
-    });
-    const { compilation } = stats;
-    const { warnings, errors } = compilation;
-
-    const transformedAsset = path.resolve(
-      __dirname,
-      outputDir,
-      "other/loader-test.webp"
-    );
-
-    const transformedExt = await fileType.fromFile(transformedAsset);
-
-    expect(/image\/webp/i.test(transformedExt.mime)).toBe(true);
-    expect(warnings).toHaveLength(0);
-    expect(errors).toHaveLength(0);
-  });
-
-  it("should emit new transformed to webp asset with filename when filename is function", async () => {
-    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-3");
-    const stats = await runWebpack({
-      entry: path.join(fixturesPath, "./loader-single.js"),
-      imageminLoaderOptions: {
-        filename: () => "other/[name].webp",
-        generator: [
-          {
-            preset: "webp",
-            implementation: ImageMinimizerPlugin.squooshGenerate,
-            options: {
-              encodeOptions: {
-                webp: {
-                  lossless: 1,
-                },
-              },
-            },
-          },
-        ],
-      },
-      output: {
-        path: outputDir,
-      },
-    });
-    const { compilation } = stats;
-    const { warnings, errors } = compilation;
-
-    const transformedAsset = path.resolve(
-      __dirname,
-      outputDir,
-      "other/loader-test.webp"
-    );
-
-    const transformedExt = await fileType.fromFile(transformedAsset);
-
-    expect(/image\/webp/i.test(transformedExt.mime)).toBe(true);
-    expect(warnings).toHaveLength(0);
-    expect(errors).toHaveLength(0);
-  });
-});
