+++ test/loader-minimizer-option.test.js
@@ -1,6 +1,8 @@
+import path from "path";
+import fileType from "file-type";
 import ImageMinimizerPlugin from "../src";
 
-import { runWebpack, isOptimized, plugins } from "./helpers";
+import { runWebpack, isOptimized, plugins, fixturesPath } from "./helpers";
 
 describe("loader minify option", () => {
   it("should work with object notation of the 'minifier' option", async () => {
@@ -409,5 +411,199 @@ describe("loader minify option", () => {
     await expect(isOptimized("loader-test.jpg", compilation)).resolves.toBe(
       true
     );
+  });
+
+  it("should generate the new webp image with flat filename", async () => {
+    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-1");
+    const stats = await runWebpack({
+      entry: path.join(fixturesPath, "./loader-single.js"),
+      imageminLoaderOptions: {
+        generator: [
+          {
+            preset: "webp",
+            implementation: ImageMinimizerPlugin.squooshGenerate,
+            filename: "[name].webp",
+            options: {
+              encodeOptions: {
+                webp: {
+                  lossless: 1,
+                },
+              },
+            },
+          },
+        ],
+      },
+      output: {
+        path: outputDir,
+      },
+    });
+
+    const { compilation } = stats;
+    const { warnings, errors } = compilation;
+
+    const transformedAsset = path.resolve(
+      __dirname,
+      outputDir,
+      "loader-test.webp"
+    );
+
+    const transformedExt = await fileType.fromFile(transformedAsset);
+
+    expect(/image\/webp/i.test(transformedExt.mime)).toBe(true);
+    expect(warnings).toHaveLength(0);
+    expect(errors).toHaveLength(0);
+  });
+
+  it("should generate the new webp image with nested filename", async () => {
+    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-2");
+    const stats = await runWebpack({
+      entry: path.join(fixturesPath, "./loader-single.js"),
+      imageminLoaderOptions: {
+        generator: [
+          {
+            preset: "webp",
+            implementation: ImageMinimizerPlugin.squooshGenerate,
+            filename: "deep/[path][name].webp",
+            options: {
+              encodeOptions: {
+                webp: {
+                  lossless: 1,
+                },
+              },
+            },
+          },
+        ],
+      },
+      output: {
+        path: outputDir,
+      },
+    });
+    const { compilation } = stats;
+    const { warnings, errors } = compilation;
+
+    const transformedAsset = path.resolve(
+      __dirname,
+      outputDir,
+      "deep/nested/deep/loader-test.webp"
+    );
+
+    const transformedExt = await fileType.fromFile(transformedAsset);
+
+    expect(/image\/webp/i.test(transformedExt.mime)).toBe(true);
+    expect(warnings).toHaveLength(0);
+    expect(errors).toHaveLength(0);
+  });
+
+  it("should generate the new webp image with filename  pointing to other directory", async () => {
+    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-3");
+    const stats = await runWebpack({
+      entry: path.join(fixturesPath, "./loader-single.js"),
+      imageminLoaderOptions: {
+        generator: [
+          {
+            preset: "webp",
+            implementation: ImageMinimizerPlugin.squooshGenerate,
+            filename: "other/[name].webp",
+            options: {
+              encodeOptions: {
+                webp: {
+                  lossless: 1,
+                },
+              },
+            },
+          },
+        ],
+      },
+      output: {
+        path: outputDir,
+      },
+    });
+    const { compilation } = stats;
+    const { warnings, errors } = compilation;
+
+    const transformedAsset = path.resolve(
+      __dirname,
+      outputDir,
+      "other/loader-test.webp"
+    );
+
+    const transformedExt = await fileType.fromFile(transformedAsset);
+
+    expect(/image\/webp/i.test(transformedExt.mime)).toBe(true);
+    expect(warnings).toHaveLength(0);
+    expect(errors).toHaveLength(0);
+  });
+
+  it("should generate the new webp image when filename is function", async () => {
+    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-4");
+    const stats = await runWebpack({
+      entry: path.join(fixturesPath, "./loader-single.js"),
+      imageminLoaderOptions: {
+        generator: [
+          {
+            preset: "webp",
+            implementation: ImageMinimizerPlugin.squooshGenerate,
+            filename: () => "other/[name].webp",
+            options: {
+              encodeOptions: {
+                webp: {
+                  lossless: 1,
+                },
+              },
+            },
+          },
+        ],
+      },
+      output: {
+        path: outputDir,
+      },
+    });
+    const { compilation } = stats;
+    const { warnings, errors } = compilation;
+
+    const transformedAsset = path.resolve(
+      __dirname,
+      outputDir,
+      "other/loader-test.webp"
+    );
+
+    const transformedExt = await fileType.fromFile(transformedAsset);
+
+    expect(/image\/webp/i.test(transformedExt.mime)).toBe(true);
+    expect(warnings).toHaveLength(0);
+    expect(errors).toHaveLength(0);
+  });
+
+  it("should minimize image with flat filename", async () => {
+    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-5");
+    const stats = await runWebpack({
+      entry: path.join(fixturesPath, "./simple.js"),
+      imageminLoaderOptions: {
+        minimizer: [
+          {
+            implementation: ImageMinimizerPlugin.squooshMinify,
+            filename: "[name][ext]",
+          },
+        ],
+      },
+      output: {
+        path: outputDir,
+      },
+    });
+
+    const { compilation } = stats;
+    const { warnings, errors } = compilation;
+
+    const transformedAsset = path.resolve(
+      __dirname,
+      outputDir,
+      "loader-test.jpg"
+    );
+
+    const transformedExt = await fileType.fromFile(transformedAsset);
+
+    expect(/image\/jpeg/i.test(transformedExt.mime)).toBe(true);
+    expect(warnings).toHaveLength(0);
+    expect(errors).toHaveLength(0);
   });
 });
